backend/requirements.txt
absl-py==1.4.0
aiohttp==3.9.5
aiosignal==1.3.1
aisuite==0.1.6
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.4.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
babel==2.16.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
betterproto==2.0.0b6
bleach==6.2.0
blinker==1.8.2
cachetools==5.3.1
category-encoders==2.6.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.1.0
colorlog==6.9.0
comm==0.1.3
contourpy==1.0.7
cryptography==44.0.0
cycler==0.11.0
databricks-sdk==0.39.0
datasets==3.2.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docker-pycreds==0.4.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
envoy==0.0.3
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.110.0
fastapi-cli==0.0.7
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==23.5.26
fonttools==4.39.4
fpdf2==2.8.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.4.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
grpcio==1.71.0
grpcio-status==1.62.3
grpclib==0.4.7
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
h5py==3.9.0
hpack==4.1.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.26.3
hyperframe==6.1.0
id==1.5.0
idna==3.4
importlib-resources==5.13.0
importlib_metadata==8.5.0
importnb==2023.11.1
in-toto-attestation==0.9.3
iniconfig==2.0.0
IProgress==0.4
ipykernel==6.23.1
ipython==8.13.2
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.5
jiter==0.8.0
joblib==1.2.0
jose==1.0.0
json5==0.9.28
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.17
kagglehub==0.3.7
keras==2.13.1
kiwisolver==1.4.4
langchain==0.3.4
langchain-core==0.3.13
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.137
libclang==16.0.6
Mako==1.3.8
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
mlflow==2.19.0
mlflow-skinny==2.19.0
model-signing==0.2.0
mongoose==0.0.1
motor==3.3.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.18.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.3
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.24.3
oauthlib==3.2.2
ollama==0.3.3
openai==0.28.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opt-einsum==3.3.0
optuna==4.2.1
orjson==3.10.10
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.0.1
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
path==17.0.0
path.py==12.5.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pid==3.0.4
pillow==11.2.1
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.0
prompt-toolkit==3.0.38
proto-plus==1.26.1
protobuf==4.25.7
psutil==5.9.5
ptyprocess==0.7.0
pub==0.0.4
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.2
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.15.1
PyJWT==2.8.0
pymongo==4.6.1
pyOpenSSL==25.0.0
pyparsing==3.0.9
PyPDF2==3.0.1
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.4.0
python-json-logger==2.0.7
python-multipart==0.0.9
python-slugify==8.0.4
pytokr==1.0
pytz==2023.3
PyYAML==6.0.2
pyzmq==25.0.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3161-client==0.1.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc8785==0.1.4
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
securesystemslib==1.2.0
selenium==4.27.1
Send2Trash==1.8.3
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
sigstore==3.6.1
sigstore-protobuf-specs==0.3.2
sigstore-rekor-types==0.0.18
six==1.16.0
sklearn==0.0.post5
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
sseclient-py==1.8.0
stack-data==0.6.2
starlette==0.36.3
statsmodels==0.14.0
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.1.0
tiktoken==0.7.0
timm==1.0.15
tinycss2==1.4.0
together==0.1.4
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.2
torch==2.6.0
torchvision==0.21.0
tornado==6.3.2
tqdm==4.66.4
traitlets==5.9.0
transformers==4.47.1
trio==0.27.0
trio-websocket==0.11.1
tuf==5.1.0
typer==0.15.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.28.0
uvloop==0.21.0
vercel==0.2.1
wandb==0.19.1
watchfiles==1.0.4
wcwidth==0.2.6
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.15.0
wsproto==1.2.0
xgboost==2.1.4
xxhash==3.5.0
yarl==1.9.4
zipp==3.21.0
